Build scripts: Skip dsvrandom if missing

+++ Rakefile
@@ -4,7 +4,11 @@ require 'zip'
 require 'pathname'
 
 require_relative 'version'
-require_relative 'dsvrandom/version'
+begin
+  require_relative 'dsvrandom/version'
+rescue LoadError
+  # Swallow
+end
 
 task :build_ui do
   # Use rbuic4 to compile the ui files into ruby code.
@@ -30,7 +34,7 @@ task :build_ui do
   system "rbuic4 dsvedit/special_object_editor.ui  -o dsvedit/ui_special_object_editor.rb"
   system "rbuic4 dsvedit/weapon_synth_editor.ui    -o dsvedit/ui_weapon_synth_editor.rb"
   
-  system "rbuic4 dsvrandom/randomizer.ui           -o dsvrandom/ui_randomizer.rb"
+  system "rbuic4 dsvrandom/randomizer.ui           -o dsvrandom/ui_randomizer.rb" if defined?(DSVRANDOM_VERSION)
 end
 
 task :build_installers do
@@ -43,13 +47,15 @@ task :build_installers do
   # OCRA normally places all the source files in the /src directory. In order to make it place them in the base directory open up /bin/ocra and change line 204 from SRCDIR = Pathname.new('src') to SRCDIR = Pathname.new('.').
 
   system "ruby ocra-1.3.6/bin/ocra dsvedit.rb --output DSVEdit.exe --no-lzma --chdir-first --innosetup setup_dsvedit.iss --icon ./images/dsvedit_icon.ico"
-  system "ruby ocra-1.3.6/bin/ocra dsvrandom/dsvrandom.rb --output DSVRandom.exe --no-lzma --chdir-first --innosetup setup_dsvrandom.iss --icon ./dsvrandom/images/dsvrandom_icon.ico"
+  system "ruby ocra-1.3.6/bin/ocra dsvrandom/dsvrandom.rb --output DSVRandom.exe --no-lzma --chdir-first --innosetup setup_dsvrandom.iss --icon ./dsvrandom/images/dsvrandom_icon.ico" if defined?(DSVRANDOM_VERSION)
 end
 
 task :build_releases do
   # Updates the executable builds with any changes to the code, delete unnecessary files, and then pack everything into zip files.
   
   ["DSVania Editor", "DSVania Randomizer"].each do |program_name|
+    next if program_name == "DSVania Randomizer" && !defined?(DSVRANDOM_VERSION)
+    
     FileUtils.rm_f ["../build/#{program_name}/armips", "../build/#{program_name}/asm", "../build/#{program_name}/constants", "../build/#{program_name}/dsvlib", "../build/#{program_name}/images", "../build/#{program_name}/dsvlib.rb", "../build/#{program_name}/version.rb"]
     FileUtils.cp_r ["./armips", "./asm", "./constants", "./dsvlib", "dsvlib.rb"], "../build/#{program_name}"
